LPS-61952 Rename com.liferay.portlet.backgroundtask to com.liferay.background.task.kernel
