+++ portal-service/src/com/liferay/background/task/kernel/util/comparator/BackgroundTaskCompletionDateComparator.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.backgroundtask.util.comparator;
+package com.liferay.background.task.kernel.util.comparator;
 
 import com.liferay.portal.kernel.backgroundtask.BackgroundTask;
 import com.liferay.portal.kernel.util.DateUtil;
@@ -21,20 +21,22 @@ import com.liferay.portal.kernel.util.OrderByComparator;
 /**
  * @author Eduardo Garcia
  */
-public class BackgroundTaskCreateDateComparator
+public class BackgroundTaskCompletionDateComparator
 	extends OrderByComparator<BackgroundTask> {
 
-	public static final String ORDER_BY_ASC = "BackgroundTask.createDate ASC";
+	public static final String ORDER_BY_ASC =
+		"BackgroundTask.completionDate ASC";
 
-	public static final String ORDER_BY_DESC = "BackgroundTask.createDate DESC";
+	public static final String ORDER_BY_DESC =
+		"BackgroundTask.completionDate DESC";
 
-	public static final String[] ORDER_BY_FIELDS = {"createDate"};
+	public static final String[] ORDER_BY_FIELDS = {"completionDate"};
 
-	public BackgroundTaskCreateDateComparator() {
+	public BackgroundTaskCompletionDateComparator() {
 		this(false);
 	}
 
-	public BackgroundTaskCreateDateComparator(boolean ascending) {
+	public BackgroundTaskCompletionDateComparator(boolean ascending) {
 		_ascending = ascending;
 	}
 
@@ -43,7 +45,8 @@ public class BackgroundTaskCreateDateComparator
 		BackgroundTask backgroundTask1, BackgroundTask backgroundTask2) {
 
 		int value = DateUtil.compareTo(
-			backgroundTask1.getCreateDate(), backgroundTask2.getCreateDate());
+			backgroundTask1.getCompletionDate(),
+			backgroundTask2.getCompletionDate());
 
 		if (_ascending) {
 			return value;
